+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## [12.8.21] - 2026-05-23
+## [12.9.0] - 2026-05-24
 
-### Added
+### Fixed
 
-- ⚔️ **AFK Journey**: Added Gwyneth and Sylphyra in preparation for patch 1.7.1.
+- ⚔️ **AFK Journey - Ravaged Realm**: Fixed the battle not starting, now correctly clicks through all wave prep screens before initiating combat.
+- ⚔️ **AFK Journey - Ravaged Realm**: Fixed "Battle over screen not found after skipping"